requirements.txt
Django==2.2.1
googleapis-common-protos==1.5.10
grpcio==1.20.1
grpcio-tools==1.20.1
image==1.5.27
Pillow==6.0.0
protobuf==3.7.1
pytz==2019.1
qrencode==1.2
six==1.12.0
sqlparse==0.3.0
